+++ src/debug.c
@@ -3157,7 +3157,7 @@ void cpu_history_init_buffer(void)
 
 }
 
-long int cpu_history_get_offset_index(int indice)
+long long int cpu_history_get_offset_index(int indice)
 {
 	return indice*CPU_HISTORY_REGISTERS_SIZE;
 }
@@ -3173,7 +3173,7 @@ void cpu_history_add_element(void)
 
 
 	//Obtener posicion en memoria
-	long int offset_memoria;
+	long long int offset_memoria;
 	offset_memoria=cpu_history_get_offset_index(cpu_history_siguiente_posicion);
 	
 	//printf ("Offset en memoria: %ld\n",offset_memoria);
@@ -3238,7 +3238,7 @@ void cpu_history_get_registers_element(int indice,char *string_destino)
 
 	int posicion=cpu_history_get_array_pos_element(indice);
 
-	long int offset_memoria=cpu_history_get_offset_index(posicion);
+	long long int offset_memoria=cpu_history_get_offset_index(posicion);
 
 	cpu_history_regs_bin_to_string(&cpu_history_memory_buffer[offset_memoria],string_destino);
 }
@@ -3258,7 +3258,7 @@ void cpu_history_get_pc_register_element(int indice,char *string_destino)
 
 	int posicion=cpu_history_get_array_pos_element(indice);
 
-	long int offset_memoria=cpu_history_get_offset_index(posicion);
+	long long int offset_memoria=cpu_history_get_offset_index(posicion);
 
 	cpu_history_reg_pc_bin_to_string(&cpu_history_memory_buffer[offset_memoria],string_destino);
 }
@@ -3283,7 +3283,7 @@ void cpu_history_regs_bin_restore(int indice)
 
 	int posicion=cpu_history_get_array_pos_element(indice);
 
-	long int offset_memoria=cpu_history_get_offset_index(posicion);
+	long long int offset_memoria=cpu_history_get_offset_index(posicion);
 
 
     z80_byte *p;
@@ -7339,7 +7339,7 @@ int remote_load_source_code(char *archivo)
 	if (remote_raw_source_code_pointer!=NULL) free(remote_raw_source_code_pointer);
 
 	//Ver tamanyo archivo
-	long int tamanyo;
+	long long int tamanyo;
 
 	tamanyo=get_file_size(archivo);
